Clean shutdown and chain height in notification
Changed the text while scanning the block chain to the height of the
block chain scanned
Added dcrd server status to enable the user monitor the status(Whether
it is loading the chain or server is already started
Dcrd now shuts down properly instead of closing without saving the block
chain and the application also closes after tapping the shutdown button
in the notification.
Chain height is displayed in dcrd service notification to enable the
user monitor the chain height without having to open the application.
Dcrwallet does not crash the app by running twice, a check is made
before starting dcrwallet to make sure no instance is already running