restore wallet: remove RecyclerView and use a custom populated layout
The RecylerView used in restore wallet activity was not the best choice
as its values in the text field gets mixed up when scrolling and sometimes
the text visible isn't the value validated. This fixes the issues by
replacing the RecyclerView with a LinearLayout and the views have their
exclusive scope rather than operating under RecyclerView and that prevents
mixup from happening between text fields.