Fix duplicate politeia notification listener error
An error that causes the app to crash occurs when multiple notification
listeners are added with the same key. This occurred because a notification
listener added previously in the app onResume was not removed in onStop
which can happen if the OS fails to call onStop. This error is avoided by
removing any notification listener added with the key before adding a new
notification listener.